Переделал ContentView по паттерн MVVM
